+++ src/key_value_store.js
@@ -290,6 +290,43 @@ export class KeyValueStore {
                 if (this.storeName) storesCache.remove(this.storeName);
             });
     }
+
+    /**
+     * Iterates over key value store keys, yielding each in turn to an `iteratee` function.
+     * Each invocation of `iteratee` is called with two arguments: `(item, index)`.
+     *
+     * If the `iteratee` function returns a Promise then it is awaited before the next call.
+     * If it throws an error, the iteration is aborted and the `forEachKey` function throws the error.
+     *
+     * **Example usage**
+     * ```javascript
+     * const keyValueStore = await Apify.openKeyValueStore();
+     * keyValueStore.forEachKey(async (key, index) => {
+     *   console.log(`Key at ${index}: ${key}`);
+     * });
+     * ```
+     *
+     * @param {Function} iteratee A function that is called for every key in the key value store.
+     * @param {Object} [options] All `forEachKey()` parameters are passed
+     *   via an options object with the following keys:
+     * @param {string} [options.exclusiveStartKey] All keys up to this one (including) are skipped from the result.
+     * @return {Promise}
+     */
+    async forEachKey(iteratee, options = {}, index = 0) {
+        const { exclusiveStartKey } = options;
+        checkParamOrThrow(iteratee, 'iteratee', 'Function');
+        checkParamOrThrow(exclusiveStartKey, 'options.exclusiveStartKey', 'Maybe String');
+        checkParamOrThrow(index, 'index', 'Number');
+
+        const response = await keyValueStores.listKeys({ storeId: this.storeId, exclusiveStartKey });
+        const { nextExclusiveStartKey, isTruncated, items } = response.data;
+        for (const key of items) {
+            await iteratee(key, index++);
+        }
+        return isTruncated
+            ? this.forEachKey(iteratee, { exclusiveStartKey: nextExclusiveStartKey }, index)
+            : undefined; // [].forEach() returns undefined.
+    }
 }
 
 /**
@@ -362,6 +399,23 @@ export class KeyValueStoreLocal {
             .then(() => {
                 storesCache.remove(this.storeId);
             });
+    }
+
+    async forEachKey(iteratee, options = {}, index = 0) {
+        const { exclusiveStartKey } = options;
+        checkParamOrThrow(iteratee, 'iteratee', 'Function');
+        checkParamOrThrow(exclusiveStartKey, 'options.exclusiveStartKey', 'Maybe String');
+        checkParamOrThrow(index, 'index', 'Number');
+
+        const files = await readdirPromised(this.localStoragePath);
+        let keys = files.map(file => path.parse(file).name).sort(); // Array is sorted to emulate API.
+        if (exclusiveStartKey) {
+            const keyPos = keys.indexOf(exclusiveStartKey);
+            if (keyPos !== -1) keys = keys.slice(keyPos + 1);
+        }
+        for (const key of keys) {
+            await iteratee(key, index++);
+        }
     }
 
     /**