+++ CHANGELOG.md
@@ -1,6 +1,8 @@
-xxx
+0.11.5 / 2018-01-18
 ===================
-- Better error message for an invalid `launchPuppeteerFunction` passed to `PuppeteerPool`
+- Disabled the `PuppeteerPool` `reusePages` option for now, due to a memory leak.
+- Added a `keyValueStore.forEachKey()` method to iterate all keys in the store.
+- Better error message for an invalid `launchPuppeteerFunction` passed to `PuppeteerPool`.
 
 0.11.4 / 2018-01-15
 ===================